ファイルに切り出し
