+++ run.py
@@ -11,7 +11,8 @@ from concurrent.futures import ThreadPoolExecutor
 from slackeventsapi import SlackEventAdapter
 import slackbot_settings as conf
 import plugins.hato as hato
-from library.clientclass import SlackClient, BaseClient
+import plugins.analyze as analyze
+from library.clientclass import SlackClient
 
 
 slack_events_adapter = SlackEventAdapter(
@@ -30,39 +31,12 @@ def __init__():
         logging.WARNING)
 
 
-def analyze_message(message: str) -> Callable[[BaseClient], None]:
-    """コマンド解析"""
-
-    conditions = {'help': lambda m: hato.help_message,
-                  'eq': lambda m: hato.earth_quake,
-                  '地震': lambda m: hato.earth_quake,
-                  'in': lambda m: hato.labotter_in,
-                  'rida': lambda m: hato.labotter_rida,
-                  'text list': lambda m: hato.get_text_list,
-                  'text add ': lambda m: hato.add_text(m[len('text add '):]),
-                  'text show ': lambda m: hato.show_text(m[len('text show '):]),
-                  'text delete ': lambda m: hato.delete_text(m[len('text delete '):]),
-                  'text random': lambda m: hato.show_random_text,
-                  'text': lambda m: hato.show_random_text,
-                  '天気': lambda m: hato.weather((m[len('天気'):]).strip()),
-                  '>< ': lambda m: hato.totuzensi(m[len('>< '):]),
-                  'amesh': lambda m: hato.amesh,
-                  'amesh ': lambda m: hato.amesh_with_gis((m[len('amesh '):]).strip()),
-                  'version': lambda m: hato.version,
-                  }
-
-    for key, method in conditions.items():
-        if message.startswith(key):
-            return method(message)
-    return hato.default_action
-
-
 def analyze_slack_message(messages: List[dict]) -> Callable[[SlackClient], None]:
     """Slackコマンド解析"""
 
     if len(messages) > 0 and messages[0]['type'] == 'text':
         message = messages[0]['text'].strip()
-        return analyze_message(message)
+        return analyze.analyze_message(message)
 
     return hato.default_action
 
@@ -88,7 +62,7 @@ def on_app_mention(event_data):
                     block_element_elements = block_element['elements']
                     if len(block_element_elements) > 0 and \
                             block_element_elements[0]['type'] == 'user' and \
-                        block_element_elements[0]['user_id'] in authed_users:
+                    block_element_elements[0]['user_id'] in authed_users:
                         TPE.submit(analyze_slack_message(block_element_elements[1:]), SlackClient(
                             channel, block_element_elements[0]['user_id']))
